Alterações na estrutura do projeto
