+++ src/App.java
@@ -1,33 +1,7 @@
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
 public class App {
 
-    public static void main(String[] args) throws Exception {
-
-        VerificaRegras regras = new VerificaRegras("regras.txt");
-        String informacao = regras.getDadosArquivo();
-
-        List<Map<String, String>> dadosPlantacoes = new ArrayList<>();
-        dadosPlantacoes = regras.getPlantacoes();
-        System.out.println("imprime dados de plantacoes: " + dadosPlantacoes);
-
-        String data = regras.getDataArquivo();
-        System.out.println(data);
-        Integer[] rangeEnchimento = new Integer[2];
-        rangeEnchimento = regras.getRangeEnchimentoSegundos();
-        Integer[] rangeRecepcao = new Integer[2];
-        rangeRecepcao = regras.getRangeRecepcao();
-        System.out.println("de " + rangeRecepcao[0] + " a " + rangeRecepcao[1] + "segundos");
-
-        System.out.println("getPlantacoes\n\t" + regras.getPlantacoes());
-        System.out.println("getCapacidadeRecepcaoLagar\n\t" + regras.getCapacidadeRecepcaoLagar());
-        System.out.print("getRangeCapacidadeCaminhao\n\t");
-        for (int i = 0; i < regras.getRangeCapacidadeCaminhao().length; i++) {
-            System.out.print(regras.getRangeCapacidadeCaminhao()[i] + " ");
-        }
-
+    public static void main(String[] args) {
+        System.out.println("Hello azeitonas!");
     }
 
 }
