--- src/Plantacao.java
@@ -1,11 +1,0 @@
-public class Plantacao {
-    Azeitonas tipoPlantacao;
-    Integer distanciaLagar;
-    Boolean produzindo;
-    Boolean aberto;
-
-    public void encherCaminhao() {
-
-    }
-
-}
