+++ src/input/regras.txt
