+++ src/plantacao/Azeitonas.java
@@ -1,6 +1,7 @@
+package plantacao;
+
 public enum Azeitonas {
     Galega,
     Cordovil,
     Picual
-
 }
