+++ src/regras/VerificaRegras.java
@@ -1,3 +1,5 @@
+package regras;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
